requirements.txt
fastapi
unicorn
autopep8
pydantic
SQLAlchemy
alembic
pymysql
python-jose[cryptography]
passlib[bcrypt]
python-dotenv